Meal-Plan-System/backend/requirements.txt
# FastAPI backend
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.0
pytest>=7.0.0

# Database
sqlalchemy>=2.0.0

# Auth
bcrypt>=4.0.0
PyJWT>=2.8.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# ML/RAG (optional - for semantic search)
scikit-learn>=1.3.0
sentence-transformers==2.2.2
chromadb==0.4.24
rapidfuzz>=3.0.0
typesense>=0.21.0
